statemachine: bound the decrease or removal of fbcoin from CF tx
We could have happily removed hundreds of thousands of sats from the fb
coins. This is hundreds of dollars and is not reasonable. It was also
not bounded and would have happily removed a lot of feebump coins. If we
end up in a situation where we need to remove that much we should not
have created the CF with these parameters in the first place!

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>